requirements.txt
python-telegram-bot==20.7
maigret>=0.4.4
httpx>=0.25.0
